Create VHF-parse workspace
This is a slightly mangled way of doing Rust workspaces, as the various
workspaces are not siblings of each other in the filesystem.
Nonetheless, this is probably more sensible.

Moved types and consts as files to be used from both packages.
