+++ VHF-rs/src/parser/consts.rs
@@ -1,1 +1,3 @@
+// XXX: This file is shared by both VHF and VHF-parse package.
+
 pub const M_OVERFLOW: u16 = 0xF000;
