+++ VHF-rs/src/parser/mod.rs
@@ -1,1 +1,2 @@
 pub mod consts;
+pub mod types;
